requirements.txt
pycryptodome==3.9.1
PyInstaller==3.5